style: 패키지 구조 정리
- 이동 조건들을 전략 패키지로 이동
- 단위 이동 ENUM을 보드 패키지로 책임위임
