+++ src/main/java/domain/piece/PieceType.java
@@ -1,5 +1,14 @@
 package domain.piece;
 
+import domain.piece.strategy.CanonMovingCondition;
+import domain.piece.strategy.ChariotMovingCondition;
+import domain.piece.strategy.CounselorMovingCondition;
+import domain.piece.strategy.ElephantMovingCondition;
+import domain.piece.strategy.GeneralMovingCondition;
+import domain.piece.strategy.HorseMovingCondition;
+import domain.piece.strategy.MovingCondition;
+import domain.piece.strategy.PawnMovingCondition;
+
 public enum PieceType {
     GENERAL("궁", new GeneralMovingCondition()),
     CHARIOT("차", new ChariotMovingCondition()),
