+++ src/main/java/domain/piece/strategy/ElephantMovingCondition.java
@@ -1,10 +1,11 @@
-package domain.piece;
+package domain.piece.strategy;
 
+import domain.piece.Piece;
 import domain.position.Position;
 
 import java.util.Map;
 
-public class HorseMovingCondition implements MovingCondition {
+public class ElephantMovingCondition implements MovingCondition {
     @Override
     public boolean canMove(Map<Position, Piece> state, Position startPosition, Position endPosition) {
         return false;
