+++ src/main/java/domain/piece/strategy/MovingCondition.java
@@ -1,5 +1,6 @@
-package domain.piece;
+package domain.piece.strategy;
 
+import domain.piece.Piece;
 import domain.position.Position;
 
 import java.util.Map;
