+++ src/main/java/domain/piece/strategy/PawnMovingCondition.java
@@ -1,5 +1,8 @@
-package domain.piece;
+package domain.piece.strategy;
 
+import domain.board.Direction;
+import domain.piece.Piece;
+import domain.piece.Side;
 import domain.position.Position;
 
 import java.util.Map;
